// apps/backend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  editor: typeof editor;
  exportEnv: typeof exportEnv;
  exportWorkpool: typeof exportWorkpool;
  exports: typeof exports;
  exportsActions: typeof exportsActions;
  http: typeof http;
  lib: typeof lib;
  projectFiles: typeof projectFiles;
  projects: typeof projects;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;